+++ apps/laboratory/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @apps/laboratory
 
+## 4.0.9
+
+### Patch Changes
+
+- Add all rpc methods + auto reject when modal closes
+
+- Updated dependencies []:
+  - @web3modal/ethers@4.0.9
+  - @web3modal/wagmi@4.0.9
+  - @web3modal/ethers5@4.0.9
+
+## 4.0.8
+
+### Patch Changes
+
+- [#1954](https://github.com/WalletConnect/web3modal/pull/1954) [`c3366e7`](https://github.com/WalletConnect/web3modal/commit/c3366e7211dba2f5c6d3377c9d9a77da5a52c0d8) Thanks [@tomiir](https://github.com/tomiir)! - Add support for eth_getBlockByNumber
+
+- Updated dependencies [[`c3366e7`](https://github.com/WalletConnect/web3modal/commit/c3366e7211dba2f5c6d3377c9d9a77da5a52c0d8)]:
+  - @web3modal/ethers@4.0.8
+  - @web3modal/ethers5@4.0.8
+  - @web3modal/wagmi@4.0.8
+
+## 4.0.8-f1845392.0
+
+### Patch Changes
+
+- [#1954](https://github.com/WalletConnect/web3modal/pull/1954) [`4755109`](https://github.com/WalletConnect/web3modal/commit/475510962a92ea9f4388db1d08c979d99da18e54) Thanks [@tomiir](https://github.com/tomiir)! - Add support for eth_getBlockByNumber
+
+- Updated dependencies [[`4755109`](https://github.com/WalletConnect/web3modal/commit/475510962a92ea9f4388db1d08c979d99da18e54)]:
+  - @web3modal/ethers@4.0.8-f1845392.0
+  - @web3modal/ethers5@4.0.8-f1845392.0
+  - @web3modal/wagmi@4.0.8-f1845392.0
+
+## 4.0.7
+
+### Patch Changes
+
+- Add eth_getBalance to list of allowed methods
+
+- Updated dependencies []:
+  - @web3modal/ethers@4.0.7
+  - @web3modal/ethers5@4.0.7
+  - @web3modal/wagmi@4.0.7
+
+## 4.0.6
+
+### Patch Changes
+
+- Email stability fixes
+
+- Updated dependencies []:
+  - @web3modal/ethers@4.0.6
+  - @web3modal/ethers5@4.0.6
+  - @web3modal/wagmi@4.0.6
+
 ## 4.0.5
 
 ### Patch Changes